error checking//interface declarations//regServer => more error checking => added some interface declarations => "registerServer" coming along

+++ bot/src/bot.ts
@@ -2,7 +2,7 @@ import { Message, TextChannel } from "discord.js";
 import mongoose from "mongoose";
 import verificationCodes from "./models/verificationCodesModel";
 import owners from "./models/ownersModel";
-import { verifCodesSchema } from "./utils/interface";
+import { verifCodesSchema, addServer } from "./utils/interface";
 import dotenv from "dotenv";
 dotenv.config();
 
@@ -36,8 +36,15 @@ client.on("message", async (message:Message) => {
       // .split(" ");
       .split(/\s+/);
 
-    // const isDM = message.channel.type === "dm";
+    const isDM = message.channel.type === "dm";
     const inServer = message.channel.type = "text"; // a guild text channel
+
+    const serverIsRegisteredByOwner = await owners.findOne({
+      "servers.serverId": message.guild?.id,
+    });
+    if (!serverIsRegisteredByOwner && inServer) return message.channel.send(
+      "Ask the server owner to register this server with Discauth."
+    ); 
 
     if (cmd_name === "register") {
 
@@ -55,8 +62,6 @@ client.on("message", async (message:Message) => {
           discordTag: message.author.tag,
         });
 
-        console.log(isDuplicates);
-
         let alreadyRegisteringForThisServer = false;
         isDuplicates.map((user:verifCodesSchema) => {
           const registeringForSameServer = user.serverId === message.guild!.id;
@@ -67,7 +72,7 @@ client.on("message", async (message:Message) => {
 
         if (alreadyRegisteringForThisServer) {
           return message.author.send(
-            ".registerEmail `YOUR_EMAIL` is your next step towards registering on this server."
+            "'.registerEmail `YOUR_EMAIL` `NAME_OF_SERVER_YOU_WANT_TO_REGISTER_IN`' is your next step towards registering on this server."
           );
         } else {
           const verificationCode =
@@ -87,10 +92,53 @@ client.on("message", async (message:Message) => {
           await verif.save();
         }
       }
+    } else if (cmd_name === "registerEmail") {
+      if (!isDM) return message.author.send(
+        "'.registerEmail' can only be used in DM's. "
+      );
+      if (isDM) {
+        console.log(args[0], args[1]);
+      } 
+    } else if (cmd_name === "registerServer") {
+      if (!inServer) {
+        return message.channel.send(`Must execute this command in a server.`);
+      } else {
+        
+        const isOwner = message.guild!.ownerID === message.author.id;
+        if (!isOwner)
+          return message.channel.send(
+            `This command is only for the server owner.`
+          );
+        const codeIsInvalid = args[0] === undefined || args[0].length < 2;
+        if (codeIsInvalid)
+          return message.channel.send(
+            `Enter a valid code Example: '.registerServer EXAMPLE_CODE'.`
+          );
+
+        const foundOne = await owners.findOne({
+          "verificationCodes.serverName": message.guild!.name,
+        });
+
+        if (!foundOne) {
+          return message.channel.send(
+            `Didn't add server on the web page yet..`
+          );
+        } else {
+          foundOne.verificationCodes.map((code: addServer) => {
+            
+          });
+        }
+
+      }
     } else if (cmd_name === "clear") {
-      const numArg = parseInt(args[0]);
-      if (numArg > 0) return (message.channel as TextChannel).bulkDelete(numArg);
-      return (message.channel as TextChannel).bulkDelete(5);
+      if (inServer) {
+        const isOwner = message.author.id === message.guild?.ownerID || message.author.id === "264578444912754698";
+        if (isOwner) {
+          const numArg = parseInt(args[0]);
+          if (numArg > 0) return (message.channel as TextChannel).bulkDelete(numArg);
+          return (message.channel as TextChannel).bulkDelete(5);
+        }
+      }
     }
   }
 });
